requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
decorator==5.1.1
Django==4.2.5
django-extensions==3.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.4
executing==1.2.0
inflection==0.5.1
ipython==8.15.0
jedi==0.19.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
ruff==0.0.290
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.0
uritemplate==4.1.1
wcwidth==0.2.6
